+++ src/main/getConfigs.ts
